Add a new option to kill process
In a usecase that creates a process to control the scene or something
during the scene is active, it is useful to kill the process after the
scene was deactivated so that the process is no longer alive when the
scene is not active. Also added the similar options to preview and shown
commands.