Close non-standard file descriptors
The obs-ffmpeg-mux process cannot be terminated while the process forked
by this plugin is running, which causes audio subsystem of OBS stalls.
The obs-ffmpeg-mux process is terminated when an EOF appears on the
pipe. However, since the forked process holds the file descriptor of the
write-end of the pipe, the EOF won't arrive to the obs-ffmpeg-mux
process.